restore the sync.Cond's fix deadlock
Basically it only ever happens if the RX path gets full! It's caused by
this section here (that adds a frame to the reassembler),
https://github.com/getlantern/multipath/blob/master/receivequeue.go#L41-L47

The issue goes something like this.

Add(f) is called, rq.tryAdd fails because the queue is full, a availableFrame
signal does not get fired so the read over at
https://github.com/getlantern/multipath/blob/master/receivequeue.go#L93 never gets moved fwd.

Now, the function goes to waitForSlot(), that waits for a availableSlot signal,
something that also depends on availableFrame being signaled, but the ability
for that to be signaled is now gone, since availableSlot is only fired at the
end of the read! and anything that could add stuff to the list (bar another add)
is now stuck on availableSlot.

So waiting on availableSlot means that availableFrame can't be signaled, so the
whole read loop locks up forever.

Amusingly (?) this bug was added to prevent overflows in the RX queue
https://github.com/getlantern/multipath/commit/1b8abb0f2708e4e8331d9f41625256525bfa2617

The fix I tested that seems to have worked (I've moved 200G in a multipath session with it so far)