Improve retransmit logic, tests shoudld now pass
These errors were exposed by the test suite, A few interesting things
were happening here.

A) The retransmit timer was unbounded upwards.

This meant that if a subflow was totally toasted, and a retransmit
happened over it. The entire connection would hang waiting for the
unlimited-ly high retransmit timer of that bad socket. This caused
the major hangs in the test suite, since it sets the delay to 1hr

B) The subflow selection logic for retransmits selected dupes

When a retransmit happens, it's critical to avoid selecting subflows
that have already been used for that frame, since multipath uses
TCP-like conn's there is no point in retransmitting the same
frame over the conn. Since it will either never get there, or it will
arrive late. The subflow selecting algro now checks against this.

---

Improvments could be made to ensure that the best subflow could
be selected for retransmit sucess to speed up recovery. But
that is out of scope of the commit right now.