Fix subtle race condition resulting in "Corrupted Buffer detected"
This happened when some subflows are a lot higher latency than the
others, this causes a retransmit to reset the state of the RXQ, and
then cause the buffer slice to be reset. causing a buffer corruption
error to be raised.

This bug was detected with bgp.tools clients, running in South Africa